MoneySavedCardView 추가

+++ Sennight/Sennight/Models/SmokingHabit.swift
@@ -33,7 +33,7 @@ struct SmokingHabit: Codable {
     let ID: Int
     let userID: Int
     let dailyCigarettes: Int
-    let cigarettePrice: Int
+    let cigarettePrice: Double
     let firstCigarette: String
     let smokingYears: Int
